Removed ALL put calls and fixed corresponding examples.
One problem with that is the following: *Client* examples are not really good examples anymore. This is because PUTs don’t come through the Master, and we end up receiving things from the receiver itself. It’s going through network, but coming from the same JVM. 

The BroadcastService is still fine though. This should be our primary example from now on.